Resize things for nicer app icon display
The OneColor theme used show_app_icon=false, while I like it set to
true. But merely setting it to true didn't look good, so I adjusted the
sizes.

IIUC, the icon size is given by the width of the menu icon, so this was
the main adjustment I made. The other adjustments were mostly cascading
from this one.